transactions api and code refactor

+++ moneyExchange/src/main/java/pl/niewadzj/moneyExchange/api/currency/records/CurrencyExternalResponse.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 @Builder
 public record CurrencyExternalResponse(String currency,
                                        String code,
-                                       double mid) {
+                                       Float mid) {
 }
